Fix Provider API error handelr
